error_service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.2.3
gunicorn==20.1.0
mysqlclient==2.2.6
python-dotenv==0.19.0

